pyproject.toml
[tool.poetry]
name = "wv-muilti-process-demo'"
version = "0.1.0"
description = "Webview Multi Process Demo"
authors = [":Barry-Thomas-Paul: Moss <4193389+Amourspirit@users.noreply.github.com>"]
repository = "https://github.com/Amourspirit/python-webview-multi-process-demo"
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.8"
pyobjc = [{version = ">=10.3.1", markers = "sys_platform == 'darwin'"}]
cefpython3 = [{version = "^66.1", markers = "sys_platform == 'win32'"}]
pywebview = [
    {extras = ["qt"], version = "^5.3.2", markers = "sys_platform != 'win32'"},
	{extras = ["cef"], version = "^5.3.2", markers = "sys_platform == 'win32'"},
    {version = "^5.3.2", markers = "sys_platform == 'darwin'"},
]
jedi = "^0.19.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
